# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SDR CRM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Workspace {
  id: string;
  name: string;
  ownerId: string;
  createdAt: string;
}

export interface CreateWorkspaceBody {
  name: string;
}

export type LeadCustomFieldValues = { [key: string]: unknown };

export interface Lead {
  id: string;
  workspaceId: string;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  notes?: string | null;
  stage: string;
  /** @nullable */
  assigneeId?: string | null;
  customFieldValues?: LeadCustomFieldValues;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadBodyCustomFieldValues = { [key: string]: unknown };

export interface CreateLeadBody {
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  notes?: string | null;
  stage?: string;
  /** @nullable */
  assigneeId?: string | null;
  customFieldValues?: CreateLeadBodyCustomFieldValues;
}

export type UpdateLeadBodyCustomFieldValues = { [key: string]: unknown };

export interface UpdateLeadBody {
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  notes?: string | null;
  stage?: string;
  /** @nullable */
  assigneeId?: string | null;
  customFieldValues?: UpdateLeadBodyCustomFieldValues;
}

export interface StageValidationError {
  error: string;
  missingFields: string[];
}

export interface Campaign {
  id: string;
  workspaceId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  context: string;
  promptInstructions: string;
  /** @nullable */
  triggerStage?: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateCampaignBody {
  name: string;
  /** @nullable */
  description?: string | null;
  context: string;
  promptInstructions: string;
  /** @nullable */
  triggerStage?: string | null;
  isActive?: boolean;
}

export interface UpdateCampaignBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  context?: string;
  promptInstructions?: string;
  /** @nullable */
  triggerStage?: string | null;
  isActive?: boolean;
}

export type GeneratedMessageStatus =
  (typeof GeneratedMessageStatus)[keyof typeof GeneratedMessageStatus];

export const GeneratedMessageStatus = {
  generated: "generated",
  sent: "sent",
} as const;

export interface GeneratedMessage {
  id: string;
  leadId: string;
  campaignId: string;
  content: string;
  status: GeneratedMessageStatus;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
}

export interface GenerateMessagesBody {
  campaignId: string;
  count?: number;
}

export interface SendMessageBody {
  messageId: string;
}

export interface FunnelStage {
  id: string;
  workspaceId: string;
  name: string;
  slug: string;
  order: number;
  /** @nullable */
  color?: string | null;
  requiredFields?: string[];
  isDefault: boolean;
}

export interface CreateFunnelStageBody {
  name: string;
  /** @nullable */
  color?: string | null;
  order?: number;
  requiredFields?: string[];
}

export interface UpdateFunnelStageBody {
  name?: string;
  /** @nullable */
  color?: string | null;
  order?: number;
  requiredFields?: string[];
}

export type CustomFieldFieldType =
  (typeof CustomFieldFieldType)[keyof typeof CustomFieldFieldType];

export const CustomFieldFieldType = {
  text: "text",
  number: "number",
  select: "select",
} as const;

export interface CustomField {
  id: string;
  workspaceId: string;
  name: string;
  slug: string;
  fieldType: CustomFieldFieldType;
  /** @nullable */
  options?: string[] | null;
  createdAt: string;
}

export type CreateCustomFieldBodyFieldType =
  (typeof CreateCustomFieldBodyFieldType)[keyof typeof CreateCustomFieldBodyFieldType];

export const CreateCustomFieldBodyFieldType = {
  text: "text",
  number: "number",
  select: "select",
} as const;

export interface CreateCustomFieldBody {
  name: string;
  fieldType: CreateCustomFieldBodyFieldType;
  /** @nullable */
  options?: string[] | null;
}

export interface StageCount {
  stage: string;
  stageName: string;
  count: number;
  /** @nullable */
  color?: string | null;
}

export interface DashboardMetrics {
  totalLeads: number;
  leadsByStage: StageCount[];
  totalCampaigns: number;
  totalMessagesGenerated: number;
  totalMessagesSent: number;
  recentLeads: Lead[];
}

export interface ActivityEvent {
  id: string;
  leadId: string;
  leadName: string;
  eventType: string;
  description: string;
  createdAt: string;
}

export type ListLeadsParams = {
  stage?: string;
  assigneeId?: string;
  search?: string;
};

export type ListLeadMessagesParams = {
  campaignId?: string;
};

export type GetActivityParams = {
  limit?: number;
};