# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SDR CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns in workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /workspaces/{workspaceId}/campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign details
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: campaignId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AI Messages ──────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/custom-fields:
    get:
      operationId: listCustomFields
      tags: [custom-fields]
      summary: List custom fields for workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Custom fields
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomField"
    post:
      operationId: createCustomField
      tags: [custom-fields]
      summary: Create a custom field
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomFieldBody"
      responses:
        "201":
          description: Created custom field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomField"
  /workspaces/{workspaceId}/custom-fields/{fieldId}:
    delete:
      operationId: deleteCustomField
      tags: [custom-fields]
      summary: Delete a custom field
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: fieldId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Dashboard ────────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard metrics for workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardMetrics"
  /workspaces/{workspaceId}/activity:
    get:
      operationId: getActivity
      tags: [dashboard]
      summary: Get recent activity log for workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        stage:
          type: string
        assigneeId:
          type: ["string", "null"]
        customFieldValues:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - stage
        - createdAt
        - updatedAt
...
    Campaign:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        context:
          type: string
        promptInstructions:
          type: string
        triggerStage:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - context
        - promptInstructions
        - isActive
        - createdAt
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        context:
          type: string
        promptInstructions:
          type: string
        triggerStage:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - context
        - promptInstructions
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        context:
          type: string
        promptInstructions:
          type: string
        triggerStage:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    CustomField:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        slug:
          type: string
        fieldType:
          type: string
          enum: [text, number, select]
        options:
          type: ["array", "null"]
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - slug
        - fieldType
        - createdAt
    CreateCustomFieldBody:
      type: object
      properties:
        name:
          type: string
        fieldType:
          type: string
          enum: [text, number, select]
        options:
          type: ["array", "null"]
          items:
            type: string
      required:
        - name
        - fieldType
    DashboardMetrics:
      type: object
      properties:
        totalLeads:
          type: integer
        leadsByStage:
          type: array
          items:
            $ref: "#/components/schemas/StageCount"
        totalCampaigns:
          type: integer
        totalMessagesGenerated:
          type: integer
        totalMessagesSent:
          type: integer
        recentLeads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
      required:
        - totalLeads
        - leadsByStage
        - totalCampaigns
        - totalMessagesGenerated
        - totalMessagesSent
        - recentLeads
    StageCount:
      type: object
      properties:
        stage:
          type: string
        stageName:
          type: string
        count:
          type: integer
        color:
          type: ["string", "null"]
      required:
        - stage
        - stageName
        - count
    ActivityEvent:
      type: object
      properties:
        id:
          type: string
        leadId:
          type: string
        leadName:
          type: string
        eventType:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - leadId
        - leadName
        - eventType
        - description
        - createdAt